% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{The constructor function will takes two vectors and alpha.
Assume that the variances in the population are the same
(this will mean a simple no conditional function)}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{alpha}{confidence interval}
}
\value{
list of the data frame
}
\description{
The constructor function will takes two vectors and alpha.
Assume that the variances in the population are the same
(this will mean a simple no conditional function)
}
