% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{An informative plot with the "Rttest" class}
\usage{
\method{plot}{Rttest}(x, pch = 21, bg = "Blue", cex = 3, ...)
}
\arguments{
\item{x}{named list of myconstr function}

\item{pch}{single character to be used as the default in plotting points}

\item{bg}{The color to be used for the background of the device region}

\item{cex}{A numerical value giving the amount by which plotting text
and symbols should be magnified relative to the default.}

\item{...}{from print function}
}
\value{
plot of the data frame
}
\description{
An informative plot with the "Rttest" class
}
