% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print the data frame using kableExtra::kable()}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{named list of myconstr function}

\item{...}{from print function}
}
\value{
a data frame in the kableExtra format
}
\description{
Print the data frame using kableExtra::kable()
}
