% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage1.R
\name{myconstr}
\alias{myconstr}
\title{Constructor for the function}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{alpha}{}
}
\value{
list of the data frame
}
\description{
Constructor for the function
}
