% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{An informative plot with the "Rttest" class}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{named list of myconstr function}

\item{...}{from ggplot function}
}
\value{
plot of the data frame
}
\description{
An informative plot with the "Rttest" class
}
